Fixed some major error handdling issues with syntax analysis, and added a more flexible lexical analyzer (allows negatives now).

+++ bis.h
@@ -140,7 +140,7 @@ struct pool_s
     /* Dynamically Bound Functions: */
     
     /* Simulated Annealing and Hill Climbing Parameters*/
-    float iterations, T, alpha, beta;
+    double iterations, T, alpha, beta;
     /* Selection Function */
     void (*select) (selected_s *);
     /* Crossover Function */
